Refactor repository generator

+++ bin/scripts/base-script.js
@@ -1,6 +1,9 @@
 import * as fs from 'node:fs/promises';
+import * as fsb from 'node:fs';
 import util from 'util';
 import { exec as execNonPromise } from 'child_process';
+import * as readline from 'readline';
+import * as events from 'events';
 
 export class BaseScript {
   settings;
@@ -10,11 +13,14 @@ export class BaseScript {
     this.settings = settings;
   }
 
-  async remplazeEntityInFile(file, search, replace){
-    const contents = await fs.readFile(process.cwd()+'/'+file, 'utf8');
-    const updated = contents.replace(search, replace)
+  async remplazeEntityInFile(file, search, replace) {
+    const contents = await fs.readFile(process.cwd() + '/' + file, 'utf8');
+    if(!search.includes('(')){
+      search = new RegExp(search, "g");
+    }
+    const updated = contents.replace(search, replace);
 
-    await fs.writeFile(process.cwd()+'/'+file, updated, 'utf-8', err2 => {
+    await fs.writeFile(process.cwd() + '/' + file, updated, 'utf-8', err2 => {
       if (err2) {
         throw Error(err2);
       }
@@ -22,32 +28,30 @@ export class BaseScript {
   }
 
   async loadContenFromFile(file) {
-    return await fs.readFile(file,  { encoding: 'utf8' });
+    return await fs.readFile(file, { encoding: 'utf8' });
   }
 
-  async copyFromArchitectureFolder(origin)
-  {
-    await fs.cp(this.path+'template/'+this.settings.codeTemplate+'/'+origin+'.stub', process.cwd()+'/src/'+origin);
+  async copyFromArchitectureFolder(origin) {
+    await fs.cp(this.path + 'template/' + this.settings.codeTemplate + '/' + origin + '.stub', process.cwd() + '/src/' + origin);
   }
 
-  async copyFolderFromArchitectureFolder(origin)
-  {
-    const files = await fs.readdir(this.path+'template/'+this.settings.codeTemplate+'/'+origin);
+  async copyFolderFromArchitectureFolder(origin) {
+    const files = await fs.readdir(this.path + 'template/' + this.settings.codeTemplate + '/' + origin);
 
     for (let index = 0; index < files.length; index++) {
-      await fs.cp(this.path+'template/'+this.settings.codeTemplate+'/'+origin+files[index], process.cwd()+'/src/'+origin+files[index].replace('.stub', ''));
+      await fs.cp(this.path + 'template/' + this.settings.codeTemplate + '/' + origin + files[index], process.cwd() + '/src/' + origin + files[index].replace('.stub', ''));
     }
   }
 
-  async insertInNewLineAfter(file, search, newContent){
-    await this.remplazeEntityInFile(file, search, search+'\n'+newContent);
+  async insertInNewLineAfter(file, search, newContent) {
+    await this.remplazeEntityInFile(file, search, search + '\n' + newContent);
   }
 
-  async insertContentAfter(file, search, newContent){
-    await this.remplazeEntityInFile(file, search, search+newContent);
+  async insertContentAfter(file, search, newContent) {
+    await this.remplazeEntityInFile(file, search, search + newContent);
   }
 
-  async execute(command){
+  async execute(command) {
     const exec = util.promisify(execNonPromise);
     await exec(command);
   }
@@ -58,22 +62,28 @@ export class BaseScript {
     }
   }
 
-  async copyFileFromArchitectureFolderAndRename(origin, entity)
-  {
-    const newFile = origin.replace('entity', this.kebabCase(entity));
-    await fs.cp(this.path+'template/clean-code/'+origin+'.stub', process.cwd()+'/src/'+newFile);
-    await this.fixCommonVars('/src/'+newFile, entity);
+  async copyFileFromArchitectureFolderAndRename(origin, entity, propierties = false) {
+    const newFile = origin.replaceAll('entity', this.kebabCase(entity));
+    await fs.cp(this.path + 'template/clean-code/' + origin + '.stub', process.cwd() + '/src/' + newFile);
+    await this.fixCommonVars('/src/' + newFile, entity, propierties);
+
+    return process.cwd() + '/src/' + newFile;
   }
 
-  async fixCommonVars(file, entity)
-  {
-    await this.remplazeEntityInFile(file, '{{pascalCase}}',  this.pascalCase(entity));
+  async fixCommonVars(file, entity, propierties = false) {
+    await this.remplazeEntityInFile(file, '{{pascalCase}}', this.pascalCase(entity));
+    await this.remplazeEntityInFile(file, '{{kebabCase}}', this.kebabCase(entity));
+    await this.remplazeEntityInFile(file, '{{camelCase}}', this.camelCase(entity));
+
+    if (propierties) {
+      await this.remplazeEntityInFile(file, '{{properties}}', propierties.join('\n\n'));
+    }
   }
 
   kebabCase(string) {
     return string.replace(/([a-z])([A-Z])/g, "$1-$2")
-    .replace(/[\s_]+/g, '-')
-    .toLowerCase();
+      .replace(/[\s_]+/g, '-')
+      .toLowerCase();
   }
 
   camelCase(str) {
@@ -87,5 +97,35 @@ export class BaseScript {
       .match(/[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g)
       .map((x) => x.charAt(0).toUpperCase() + x.slice(1).toLowerCase())
       .join("");
+  }
+
+  async loadEntityProperties(enitty){
+    const lines = await this.loadContendFromFileLineByLine(process.cwd()+'/src/entities/'+this.kebabCase(enitty)+'.model.ts')
+
+    return lines.filter(line => line.includes(':'))
+  }
+
+  async loadContendFromFileLineByLine(filePath) {
+    const lines = [];
+    if (!fsb.existsSync(filePath)) {
+      throw new Error('La entidad no existe');
+    }
+
+    const fileStream = fsb.createReadStream(filePath);
+
+    const file = readline.createInterface({
+      input: fileStream,
+      output: process.stdout,
+      terminal: false
+    });
+
+    file.on('line', (line) => {
+      lines.push(line);
+    });
+
+    await events.once(file, 'close');
+
+
+    return lines;
   }
 }
